lib/flagutil/bytes.go: properly handle values bigger than 2GiB on 32-bit architectures
This fixes handling of values bigger than 2GiB for the following command-line flags:

- -storage.minFreeDiskSpaceBytes
- -remoteWrite.maxDiskUsagePerURL

+++ lib/flagutil/bytes_test.go
@@ -24,7 +24,7 @@ func TestBytesSetFailure(t *testing.T) {
 }
 
 func TestBytesSetSuccess(t *testing.T) {
-	f := func(value string, expectedResult int) {
+	f := func(value string, expectedResult int64) {
 		t.Helper()
 		var b Bytes
 		if err := b.Set(value); err != nil {
